Tweak slow migration
Smaller batches seemed to help when running this in a console.

Don't overwrite any non-NULL associated records that might have been set.

Add the index earlier. This may slow down writes, but may also speed up the finding part of those updates, and means we get the benefits of the index sooner.